heartbeat/simple_heart: touch the strategy immediately on boot
I noticed this when MONITORing some stuff on Notifyd today. Previously the
strategy would only insert the key after the interval had passed. This meant
that if the daemon crashed in that period of time, any items it read
from the queue would be irretreivably lost. This PR modifies the heart
beat so that it touches the hash once before returning. This'll cause
workers to exist in that hash before they start clamoring for queue
items.